pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["LJlkdskdjflsa <liliangjya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.6"
path = "^16.4.0"
Django = "^4.1"
graphene-django = "^2.15.0"
pandas = "^1.4.4"
torch = "^1.12.1"
django-q = "^1.3.9"
tqdm = "^4.64.0"
django-metamask-auth = "^0.1"
django-import-export = "^2.8.0"
ciso8601 = "^2.2.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
